Refactor hashbrown impl as macro to reduce duplication re #628
