+++ rkyv/src/impls/ext/mod.rs
@@ -10,12 +10,12 @@
 mod arrayvec_0_7;
 #[cfg(feature = "bytes-1")]
 mod bytes_1;
-#[cfg(feature = "hashbrown-0_14")]
-mod hashbrown_0_14;
-#[cfg(feature = "hashbrown-0_15")]
-mod hashbrown_0_15;
-#[cfg(feature = "hashbrown-0_16")]
-mod hashbrown_0_16;
+#[cfg(any(
+    feature = "hashbrown-0_14",
+    feature = "hashbrown-0_15",
+    feature = "hashbrown-0_16"
+))]
+mod hashbrown;
 #[cfg(feature = "indexmap-2")]
 mod indexmap_2;
 #[cfg(feature = "smallvec-1")]
